+++ R/autoplot-randomwalk.R
@@ -118,10 +118,18 @@ tidy_random_walk_autoplot <- function(.data, .line_size = 1, .geom_rug = FALSE,
                    "Scale: ", atb$all$.scale)
         } else if (atb$all$tibble_type == "tidy_inverse_exponential"){
             paste0("Rate: ", atb$all$.rate, " - Scale: ", atb$all$.scale)
-        }else if (atb$all$tibble_type == "tidy_inverse_gaussian"){
+        } else if (atb$all$tibble_type == "tidy_inverse_gaussian"){
             paste0("Mean: ", atb$all$.mean, " - ",
                    "Shape: ", atb$all$.shape, " - ",
                    "Dispersion: ", atb$all$.dispersion)
+        } else if (atb$all$tibble_type == "tidy_generalized_beta"){
+            paste0(
+                "Shape1: ", atb$all$.shape1, " - ",
+                "Shape2: ", atb$all$.shape2, " - ",
+                "Shape3: ", atb$all$.shape3, " - ",
+                "Scale: ", atb$all$.scale, " - ",
+                "Rate: ", atb$all$.rate
+            )
         }
     )